Convert size estimates to ints

+++ src/main/java/org/opensearch/knn/common/KNNConstants.java
@@ -49,7 +49,7 @@ public class KNNConstants {
     public static final String MODEL_INDEX_NAME = ".opensearch-knn-models";
     public static final String PLUGIN_NAME = "knn";
     public static final String MODEL_METADATA_FIELD = "knn-models";
-    public static final Long BYTES_PER_KILOBYTES = 1024L;
+    public static final Integer BYTES_PER_KILOBYTES = 1024;
     public static final String PREFERENCE_PARAMETER = "preference";
 
     public static final String MODEL_STATE = "state";
